Maven directory restructuring
